06.01.2024 - удаление файлов main
